clarify: Simulation-Based Inference for Regression Models

Performs simulation-based inference as an alternative to the delta method for obtaining valid confidence intervals and p-values for regression post-estimation quantities, such as average marginal effects and predictions at representative values. This framework for simulation-based inference is especially useful when the resulting quantity is not normally distributed and the delta method approximation fails. The methodology is described in King, Tomz, and Wittenberg (2000) doi:10.2307/2669316 . 'clarify' is meant to replace some of the functionality of the archived package 'Zelig'; see the vignette "Translating Zelig to clarify" for replicating this functionality.

See also

Author

Maintainer: Noah Greifer ngreifer@iq.harvard.edu (ORCID)

Authors:

Steven Worthington sworthington@iq.harvard.edu (ORCID)

Stefano Iacus siacus@iq.harvard.edu (ORCID)

Gary King king@harvard.edu (ORCID)